Refactor routing feature to use <Layout> component in React JS
