+++ src/Components/Contact.js
@@ -1,22 +1,20 @@
 import React from "react";
-import Navbar from "./Navbar";
 import Footer from "./Footer";
 
 const Contact = () => {
   return (
     <div>
-      <Navbar />
       <div className="contact-page-wrapper">
-      <h1 className="primary-heading">Have Question In Mind?</h1>
-      <h1 className="primary-heading">Let Us Help You</h1>
-      <div className="contact-form-container">
-        <input type="text" placeholder="yourmail@gmail.com" />
-        <button className="secondary-button">Submit</button>
+        <h1 className="primary-heading">Have Question In Mind?</h1>
+        <h1 className="primary-heading">Let Us Help You</h1>
+        <div className="contact-form-container">
+          <input type="text" placeholder="yourmail@gmail.com" />
+          <button className="secondary-button">Submit</button>
+        </div>
       </div>
+      <Footer />
     </div>
-    <Footer />
-    </div>
-    
+
   );
 };
