+++ src/Components/Work.js
@@ -2,7 +2,6 @@ import React from "react";
 import PickMeals from "../Assets/pick-meals-image.png";
 import ChooseMeals from "../Assets/choose-image.png";
 import DeliveryMeals from "../Assets/delivery-image.png";
-import Navbar from "./Navbar";
 import Footer from "./Footer";
 
 const Work = () => {
@@ -26,7 +25,6 @@ const Work = () => {
 
   return (
     <div>
-      <Navbar />
       <div className="work-section-wrapper">
         <div className="work-section-top">
           {/* <p className="primary-subheading">Work</p> */}
